+++ src/entries/popup/pages/home/Points/PointsOnboardingSheet.tsx
@@ -31,119 +31,17 @@ import { zIndexes } from '~/entries/popup/utils/zIndexes';
 import * as wallet from '../../../handlers/wallet';
 
 import {
-  CATEGORY_DISPLAY_TYPE,
   CATEGORY_TYPE,
-  EXISTING_USER_ERROR,
-  INVALID_REFERRAL_CODE_ERROR,
   USER_POINTS_CATEGORY,
   USER_POINTS_ONBOARDING,
 } from './references';
 import { usePointsChallenge } from './usePointsChallenge';
+import { getDelayForRow, getDelayForRows, getErrorString } from './utils';
 
 const fadeVariants = {
   initial: { opacity: 0 },
   animate: { opacity: 1, transition: { delay: 0.5 } },
   exit: { opacity: 0, transition: { delay: 0.5 } },
-};
-
-const getErrorString = (error: string) => {
-  switch (error) {
-    case EXISTING_USER_ERROR:
-      return i18n.t('points.error.existing_user');
-    case INVALID_REFERRAL_CODE_ERROR:
-      return i18n.t('points.error.invalid_referral_code');
-    default:
-      return '';
-  }
-};
-
-const characterTypingTime = 0.1;
-
-const getDelayForRow = (rows: string[], row: number) => {
-  const characters = rows.reduce((total, str, index) => {
-    if (index <= row) {
-      return total + str.length;
-    }
-    return total;
-  }, 0);
-  return characters * characterTypingTime;
-};
-
-const getDelayForRows = (
-  rows: string[][],
-  row: number,
-  column: number,
-): number => {
-  console.log('rows.flat()', rows.flat());
-  const delay = getDelayForRow(rows.flat(), row * 2 + column - 1);
-  console.log(row, column, row * 2 + column - 1, delay);
-  return delay;
-};
-
-const DUMMY_USER = {
-  categories: [
-    {
-      data: {
-        owned_collections: 0,
-        total_collections: 0,
-        usd_amount: 1,
-      },
-      type: 'metamask-swaps' as CATEGORY_TYPE,
-      display_type: 'USD_AMOUNT' as CATEGORY_DISPLAY_TYPE,
-      earnings: { total: 0 },
-    },
-    {
-      data: {
-        owned_collections: 0,
-        total_collections: 0,
-        usd_amount: 1,
-      },
-      type: 'rainbow-bridges' as CATEGORY_TYPE,
-      display_type: 'USD_AMOUNT' as CATEGORY_DISPLAY_TYPE,
-      earnings: { total: 0 },
-    },
-    {
-      data: {
-        owned_collections: 1,
-        total_collections: 10,
-        usd_amount: 1,
-      },
-      type: 'nft-collections' as CATEGORY_TYPE,
-      display_type: 'NFT_COLLECTION' as CATEGORY_DISPLAY_TYPE,
-      earnings: { total: 0 },
-    },
-    {
-      data: {
-        owned_collections: 0,
-        total_collections: 0,
-        usd_amount: 1,
-      },
-      type: 'historic-balance' as CATEGORY_TYPE,
-      display_type: 'USD_AMOUNT' as CATEGORY_DISPLAY_TYPE,
-      earnings: { total: 0 },
-    },
-    {
-      data: {
-        owned_collections: 0,
-        total_collections: 0,
-        usd_amount: 1,
-      },
-      type: 'bonus' as CATEGORY_TYPE,
-      display_type: 'BONUS' as CATEGORY_DISPLAY_TYPE,
-      earnings: { total: 1 },
-    },
-    {
-      data: {
-        owned_collections: 0,
-        total_collections: 0,
-        usd_amount: 1,
-      },
-      type: 'rainbow-swaps' as CATEGORY_TYPE,
-      display_type: 'USD_AMOUNT' as CATEGORY_DISPLAY_TYPE,
-      earnings: { total: 0 },
-    },
-  ],
-  earnings: { total: 100 },
 };
 
 export const PointsOnboardingSheet = () => {
@@ -155,7 +53,7 @@ export const PointsOnboardingSheet = () => {
   const [accessGranted, setAccessGranted] = useState(false);
   const [error, setError] = useState<null | string>();
   const [userOnboarding, setUserOnboarding] =
-    useState<USER_POINTS_ONBOARDING>(DUMMY_USER);
+    useState<USER_POINTS_ONBOARDING>();
   const debouncedAccessGranted = useDebounce(accessGranted, 1000);
 
   const userOnboardingCategories = useMemo(() => {
@@ -570,14 +468,14 @@ export const PointsOnboardingSheet = () => {
                 </Text>
               </Inline>
 
-              {debouncedAccessGranted && (
+              {!debouncedAccessGranted && (
                 <AnimatedTextRows
                   id="animated-loading-rows"
                   rows={consoleLoadingRows}
                   rowsText={loadingRowsText}
                 />
               )}
-              {!debouncedAccessGranted && (
+              {debouncedAccessGranted && (
                 <AnimatedTextRows
                   id="animated-calculating-rows"
                   rowsText={calculatingPointsRowsText.flat()}
